refactor(api): add brand filter and split large file
Added a filter to load data only for the specified brand.
Split a large file into two smaller files for improved readability and code organization.
